NavBar: Fix Header icons moving when a Layer is opened
Due to the way a Grommet Header component is spaced, opening up a
layer within that component causes it to move everything else out of
place. Use the context to move some of the open/close logic of the
NavBar components within themselves and only keep parts that really
use the context within the wrapper.

Signed-off-by: notcarbide <luis@0k.cx>